+++ lib/graphql_operations/storefront/mutations/cart/cart_discount_code_update_mutation.dart
@@ -77,6 +77,12 @@ mutation cartDiscountCodesUpdate($cartId: ID!, $discountCodes: [String!], $count
           node {
             id
             quantity
+            discountAllocations {
+              discountedAmount {
+                amount
+                currencyCode
+              }
+            }
             sellingPlanAllocation {
               sellingPlan {
                 id